requirements.txt
streamlit==1.37.0
sentence-transformers==3.0.1
pysqlite3-binary==0.5.3
chromadb==0.5.0
numpy==1.23.4
pytest-cov==5.0.0